Display private info and allow purchasing
This isn't "smart" you just choose the person and then allow them
to buy. It reduces thier capital by the cost of the private but
that's it. You can then allow someone else to buy it, at which point
they reduce this capital by the cost but nothing else happens.